+++ changelog.md
@@ -10,12 +10,12 @@
 - [0.1.0](#010) • [0.1.1](#011) • [0.1.2](#012)
 
 ## Next
-
 - **Fixes**
   - [Syntax] Fixed keywords `@var` and `@endvar` not being parsed properly when not separated from other content by newlines.
   - [Syntax] Fixed superfluous whitespace being added to CSS output.
   - [Syntax] Fixed parser option `DECIMAL_PLACES` not being applied to all applicable values.
   - [Syntax] Fixed at-rules being prefixed with erroneous ampersands.
+  - [Scripting] Fixed custom options not being applied.
 
 ## 1.0.0
 *2021-06-19*